Update column-select tests.
