+++ src/shared/components/column-select/__tests__/ColumnSelect.spec.tsx
@@ -2,15 +2,12 @@ import React from 'react';
 import { fireEvent, waitFor } from '@testing-library/react';
 
 import ColumnSelect from '../ColumnSelect';
-import { removeFieldFromFieldsData } from '../utils';
 
 import renderWithRedux from '../../../__test-helpers__/RenderWithRedux';
 
 import { Namespace } from '../../../types/namespaces';
 import { UniProtKBColumn } from '../../../../uniprotkb/types/columnTypes';
-import { ColumnSelectTab } from '../../../../uniprotkb/types/resultsTypes';
 
-import structuredResultFieldsData from './__mocks__/structuredResultFieldsData.json';
 import resultFields from '../../../../uniprotkb/__mocks__/resultFields.json';
 import '../../../../uniprotkb/components/__mocks__/mockApi';
 
@@ -71,43 +68,5 @@ describe('ColumnSelect component', () => {
     expect(onChange).toHaveBeenCalledWith(
       selectedColumns.filter((c) => c !== UniProtKBColumn.proteinExistence)
     );
-  });
-});
-
-describe('removeFieldFromFieldsData', () => {
-  test('should remove field', () => {
-    const entryField = {
-      tabId: ColumnSelectTab.data,
-      accordionId: 'Names & Taxonomy',
-      itemId: UniProtKBColumn.accession,
-    };
-    expect(
-      removeFieldFromFieldsData(entryField, structuredResultFieldsData)
-    ).toEqual({
-      [ColumnSelectTab.data]: [
-        {
-          id: 'Names & Taxonomy',
-          title: 'title',
-          items: [
-            {
-              id: UniProtKBColumn.ccAllergen,
-              label: 'ccAllergen-label',
-            },
-          ],
-        },
-      ],
-      [ColumnSelectTab.links]: [
-        {
-          id: 'Sequence',
-          title: 'title',
-          items: [
-            {
-              id: UniProtKBColumn.xrefAbcd,
-              label: 'xrefAbcd-label',
-            },
-          ],
-        },
-      ],
-    });
   });
 });
