--- src/shared/components/column-select/__tests__/ColumnSelectView.spec.tsx
@@ -1,43 +1,0 @@
-import React from 'react';
-import { render, fireEvent } from '@testing-library/react';
-import ColumnSelectView from '../ColumnSelectView';
-import { UniProtKBColumn } from '../../../types/columnTypes';
-import structuredResultFieldsData from './__mocks__/structuredResultFieldsData.json';
-
-describe('ColumnSelectView component', () => {
-  let props;
-  let rendered;
-  beforeEach(() => {
-    props = {
-      selectedColumns: [
-        UniProtKBColumn.ccAllergen,
-        UniProtKBColumn.proteinExistence,
-      ],
-      fieldData: structuredResultFieldsData,
-      onSelect: jest.fn(),
-      onDragDrop: jest.fn(),
-      onReset: jest.fn(),
-    };
-    rendered = render(<ColumnSelectView {...props} />);
-  });
-
-  test('should render', () => {
-    const { asFragment } = rendered;
-    expect(asFragment()).toMatchSnapshot();
-  });
-
-  test('should call onSelect when an item is selected in the accordion', () => {
-    const { queryAllByTestId } = rendered;
-    const content = queryAllByTestId('accordion-content');
-    const listItemCheckbox = content[0].querySelector('li>label>input');
-    fireEvent.click(listItemCheckbox);
-    expect(props.onSelect).toHaveBeenCalledWith(UniProtKBColumn.accession);
-  });
-
-  test('should call onReset but reset to default button is clicked', () => {
-    const { getByTestId } = rendered;
-    const button = getByTestId('column-select-reset-button');
-    fireEvent.click(button);
-    expect(props.onReset).toHaveBeenCalled();
-  });
-});
