+++ src/shared/components/column-select/utils/__tests__/__mocks__/structuredResultFieldsData.json
